OPNFLWPLUG-1049 : device connection hold time yang changes to introduce hold time between consecutive switch connections
Device connection hold time is the least time delay any device has to 
mantain between its consecutive connection attempts. If time delay
between the previous connection and the current connection is within
device connection hold time, the device will not be allowed to connect
to the controller. Default value of the device connection hold time is
90 seconds

Change-Id: I8acaa5d78e5f7c5e19b3c1234200e0804cba745a
Signed-off-by: Somashekhar Javalagi <somashekhar.manohara.javalagi@ericsson.com>