WIP: optimize QName stream encoding
Profiling is showing we are spending about 46% in readQName(),
which further breaks down to 30% in readCodedString() and 61%
in QNameFactory.create().

Out of 1.7M invocations of readQName(), we end up loading only
508 QNames -- which is understandable, as the number of QNames is
bound by the SchemaContext contents.

This patch updates the streaming format to use linear coding, same
as we are doing for Strings. This has the effect of making the
stream smaller (reused QName is 5 bytes instead of 15), as well
as taking hash-based lookups completely out of the picture.

// FIXME: qualify performance improvement
// FIXME: split into read- and write-part

Change-Id: I0eb9958d7e3c56aa3eaf125c3fb8eb5bc4dd3800
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>