Leader should always apply modifications as local
After a leader -> follower -> leader transition which happens
before gaining consensus but after sending CommitTransactionPayload off
to persistence, the leader looses its ClientRequestTracker which
makes it apply the modification as foreign once it resumes leadership.

The consequence of applying the modification as foreign is that the
transaction becomes stuck forever in COMMIT_PENDING state, since it will
be never aborted or finished.

Change this so that the leader will apply modifications as local even
when it looses its ClientRequestTracker and add Identifiable to payloads
which require it.

Reproducing test included.

Change-Id: I636f998cd62ec82ef02193261624e4a51275fb86
Signed-off-by: Tomas Cere <tomas.cere@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>