WIP: Add Neon SR3 coding format
When we are emitting a leaf which is also a key in a list, the
leaf value is already captured in NodeIdentifierWithPredicates,
which has already been emitted when the leaf is serialized.

This means we can easily skip value serialization if we determine
a leaf to be part of the key. This should be an invariant and
always true -- otherwise we would be having inconsistent objects
in the system.

// FIXME: tests need to be added

JIRA: CONTROLLER-1909
Change-Id: I7d7cd56d5673c3af49e2717eb4bea6e8364ba2d1
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>