WIP: split up transaction chunks
This adds the infrastructure to prevent allocation of large byte[]s,
as when those exceed 0.5-16MiB, under G1GC they end up in humongous
object region. Not only that, but after some cut-off point, the
copying of arrays starts to dominate performance.

What we do here is ensure we always receive up to a configurable
number of bytes, defaults to 256KiB, and keep those chunks in a list.

This way we may end up with larger overhead, but that really is
neglibeble -- even a 2GiB payload would end up using only about
8K arrays.

// FIXME: optimize output implementation:
//        - retain previously-used small chunks for repetetive use:
//          - if we start with 512 and we go to 1024, keep the 512
//            and when we reach max size, store the full chunk and
//            start fillin up the retained 512, then 1024, etc.

Change-Id: I2b79a633ebf4fdf8d68d2accc644326e30b41f22
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>