Update patch set 1
Patch Set 1: Code-Review-2

I already pushed/merged a similar patch using a timer. We had discussed on the clustering call using the count approach but we thought the timer approach was cleaner as it keeps the functionality local to the Follower. This patch modifies the ElectionTimeout which is also used by Candidate.

Patch-set: 1