Update patch set 2
Patch Set 2:

In order to separate the two, the shard will need to send the PersistData message. This will occur for every transaction and thus will add another message queuing hop and context switch which will degrade throughput performance. We spent a lot of time in Li eliminating message hops, which we we found were the major performance detractor. 

While separating the two is an interesting idea, it will add more messaging between the two which is OK as long as it's not on the fast path and we don't regress in performance. But I'm afraid we will regress with the introduction of the PersistData message.

I think we need to step back and look at what we think we will buy by separating the two and is it compelling enough to potentially regress in performance. .

Patch-set: 2