Update patch set 15
Patch Set 15:

>Is the CLI bundle generation optional? 

As it is now, no it's not optional.

> If not can it be made so? 

Absolutely, we will have to look this up though.

> I don't think many/most folks will write karaf CLI commands for their apps so generating a skeleton will add noise

I tend to disagree with this, as we had many requests for karaf CLIs. Moreover, this was done as per as a feedback gathered from one TWS session (Anil was actually part of this feedback). 

> also some folks may not even know what it's for.

I hear you, this should be advertise correctly through a wiki page.

We could maybe trigger a mail thread on this in order to gather more feedback in whether or not this is of global interest. Else it can just be moved to a wiki page explaining how to do it.

Patch-set: 15