Update patch set 2
Patch Set 2: Code-Review+1

Thx Tom for clarification!
Delaying the processing of the messages till the module schema context is available can be an approach. Seems akka provides facility to stash messages and control the automatic recovery. 

http://doc.akka.io/docs/akka/snapshot/scala/persistence.html

wherein instead of using automatic recovery -- you tell the system when to recover --AFAIK there is facility of stashing messages too. 

Something to consider -- as introducing another persistence dependency to fix one might cause issues that we haven't thought of. 

For now +1 with above comments for consideration

Patch-set: 2