Update patch set 6
Patch Set 6: Code-Review+1

(2 comments)

Few things to check and apply as required (this is per my understanding of TransactionChaining)..

What happens to the transaction chain when a transaction is aborted on a transaction chain? 

What happens when a transaction on a transaction chain results in optimized lock exception ? 

Where are the checks that no new transaction can be added when a transaction chain is already in ready state?

Patch-set: 6