Update patch set 6
Patch Set 6:

Ed, are you suggesting that there is really no way to separate to smaller changes this almost 4k lines patch?
(Not to mention that client and server can and should be tested separately by unitests.)
Also, it seems this patch hasn't been rebased for a long time.
It contains the older jackson library.
Please rebase it to comply with the latest master.

Patch-set: 6