Allow persistence to be enabled/disabled on per-shard basis
Previously the persistence of all shards was dependent on the
persistence of the datastore. This patch adds the option to specify the
persistence per-shard and per-datastore. For module-based shards it can
be specified by adding container "persistence" into module-shards.conf
(see /sal-clustering-config/src/main/resources/initial/module-shards.conf)
For prefix-based shards the container "persistence" was added into
module prefix-shard-configuration. In both cases the attributes
"datastore-type" and "persistent" have to be specified:
 - datastore-type can be set to "configuration", "operational" or "both"
 - persistent is set to true or false
If no persistence is configured the persistence setting of the datastore
is used

JIRA: CONTROLLER-1916
Change-Id: Ie34605eece41e9a36bf1a91f0536cc019c1891e7
Signed-off-by: Tibor Král <tibor.kral@pantheon.tech>