Fix "Leader should always apply modifications as local" regression
The PreLeader state is attempting to commit transactions from
the previous leader which haven't yet been committed.
This means that the transaction is not yet applied to the DataTree,
hence all of them need to go through the full commit process,
otherwise they would be silently ignored and lost.

JIRA: CONTROLLER-1927
JIRA: CONTROLLER-1928
Change-Id: I2aeef337fca3765ee09caaf0bdcdeabdeefcf80d
Signed-off-by: Tomas Cere <tomas.cere@pantheon.tech>