CONTROLLER-1930: EOS fix to reset owner on the leader when the follower starts up
If due to any reason the entity ownership data
for an entity is altered to have no candidates
but an owner, the entity ownership does not recover
from such a condition in spite of restarts. On
startup, the follower requests the leader to
remove itself from the candidate. But it did
not change the owner. This fix now resets the
owner so that later when the follower
registers the entity, leader can elect a new
owner and notify the follower, in turn
initializing it.

JIRA: CONTROLLER-1930

Signed-off-by: Tejas Nevrekar <tejas.nevrekar@gmail.com>
Change-Id: Ia659db44e125d8fa8dd70d862cb87f5d17e8ab06