Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Hi Robert,
>    We are planning to remove all MDSAL deprecated APIs and keeping new APIs in "netvirt" project. For that we need an confirmation from you to start this exercise. Is it only this controller project review alone pending or should we wait for some more time.

Hey Karthik,

I expect this will be the only patch for removal. It currently covers all the removed APIs as far as I can tell, but needs some more work on test components.

I expect it to be be finished in about a week or two, and we'll hold off merging it until you are ready or the MRI integration window (4/11?) is upon us.

Patch-set: 4