Update patch set 2
Patch Set 2: Code-Review-1

If the intent for this is to have a human readable DataTree export, I don't think we need to have this in the controller project. There is the daexim project which already does export of the DataTree with json encoding.

Patch-set: 2
Reviewer: Gerrit User 3127 <3127@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1