Clean up features-config-netty
This feature repository contains a single feature, odl-config-netty,
which really is a remnant of config subsystem.

It further attempts to unify configuration of various southbound
modules' use of Netty, which unfortunately does not work because
downstreams need have better control over verious aspects of Netty
configuration.

Furthermore not everyone is using this infrastructure, and even
then it is a bad idea, as it lowers the amount of isolation plugins
get.

Repurpose this feature repository as features-controller-experimental,
holding a single odl-controller-exp-netty-config feature. This
clearly indicates the amount of stability that can be expected
and the plan is to move downstreams away from this feature.

Change-Id: I2edb9da48795cf1ff7de68c943c1bf895f584577
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>