WIP: initial proposal for AskTimeoutException alarms
// Functional spec is:
Implementation to get the ask timeout exception count.

Keep a tracker to maintain the counter and dump the counter value
to a history file.

After dumping the counter value then clear the counter and reset
it to 0.

The timer interval to reset the counter is configured via
JMX / JConsole interface.

Raising Alarms based on the Datastore Transaction Error Count.
Implemented a MBean for alarm notification.

// end of functional spec.

The implementation requires a major rework as to be properly
structured, for example not hard-coding shard names.

This will impact the functional spec, as parts of it are not
sensible here, for example the history should be output to\
logging and logging should be correctly configured to have a
history log file

First round of review:
Address most static wiring atrocities, as while we have a global
instance, it has instance methods and while one is fair enough (for
now), the other is not.

Second round of review:
Integrate DatastoreExceptionCountMXBeanImpl into tracker, they
are tightly integrated. This eliminates method name mismatches.

Third round of review:
Eliminate static analysis suppressions and fixup the issues reported
by SA.

JIRA: CONTROLLER-1963
Change-Id: I0daff7e0533be085e37ddebbb16e029a032b66d1
Signed-off-by: Sunan Nallasamy <sunan.n@altencalsoftlabs.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>